requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0

django-phonenumber-field==2.0.1
model_mommy>=1.6.0,<1.7.0

flake8>=3.6.0,<3.7.0